python-service/requirements.txt
paho-mqtt>=2.0.0
pandas
influxdb-client
